Rename `lm_eval.logging -> lm_eval.loggers` (#1858)
* rename lm_eval.logging module

* fix evaluation tracker args
